+++ stubs/gtk4.php
@@ -1862,6 +1862,16 @@ final class GKeyFile
         unset($key);
         unset($value);
     }
+    /**
+     * Associates a list of boolean values with $key under $group_name. If $key cannot be found
+     * then it is created. If $group_name is `null`, the start_group is used.
+     */
+    public function set_boolean_list(string $group_name, string $key, array $list): void
+    {
+        unset($group_name);
+        unset($key);
+        unset($list);
+    }
     /** Places a comment above $key from $group_name. */
     public function set_comment(?string $group_name, ?string $key, string $comment): bool
     {
@@ -1879,6 +1889,16 @@ final class GKeyFile
         unset($group_name);
         unset($key);
         unset($value);
+    }
+    /**
+     * Associates a list of double values with $key under $group_name. If $key cannot be found then
+     * it is created.
+     */
+    public function set_double_list(string $group_name, string $key, array $list): void
+    {
+        unset($group_name);
+        unset($key);
+        unset($list);
     }
     /**
      * Associates a new integer value with $key under $group_name. If $key cannot be found then it
@@ -1901,6 +1921,16 @@ final class GKeyFile
         unset($value);
     }
     /**
+     * Associates a list of integer values with $key under $group_name. If $key cannot be found
+     * then it is created.
+     */
+    public function set_integer_list(string $group_name, string $key, array $list): void
+    {
+        unset($group_name);
+        unset($key);
+        unset($list);
+    }
+    /**
      * Sets the character which is used to separate values in lists. Typically ';' or ',' are used
      * as separators. The default list separator is ';'.
      */
@@ -1920,6 +1950,17 @@ final class GKeyFile
         unset($string);
     }
     /**
+     * Associates a list of string values for $key and $locale under $group_name. If the
+     * translation for $key cannot be found then it is created.
+     */
+    public function set_locale_string_list(string $group_name, string $key, string $locale, array $list): void
+    {
+        unset($group_name);
+        unset($key);
+        unset($locale);
+        unset($list);
+    }
+    /**
      * Associates a new string value with $key under $group_name. If $key cannot be found then it
      * is created. If $group_name cannot be found then it is created. Unlike
      * g_key_file_set_value(), this function handles characters that need escaping, such as
@@ -1930,6 +1971,16 @@ final class GKeyFile
         unset($group_name);
         unset($key);
         unset($string);
+    }
+    /**
+     * Associates a list of string values for $key under $group_name. If $key cannot be found then
+     * it is created. If $group_name cannot be found then it is created.
+     */
+    public function set_string_list(string $group_name, string $key, array $list): void
+    {
+        unset($group_name);
+        unset($key);
+        unset($list);
     }
     /**
      * Associates a new integer value with $key under $group_name. If $key cannot be found then it
@@ -6363,6 +6414,12 @@ class GThemedIcon extends GObject implements GIcon
     public function __construct(string $iconname)
     {
         unset($iconname);
+    }
+    /** Creates a new themed icon for $iconnames. */
+    public static function new_from_names(array $iconnames): GThemedIcon
+    {
+        unset($iconnames);
+        return null;
     }
     /**
      * Creates a new themed icon for $iconname, and all the names that can be created by shortening